requirements.txt
asgiref==3.8.1
Django==5.1.7
django-simple-history==3.8.0
Faker==37.1.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
